+++ tests/testthat/test-predict.R
@@ -23,6 +23,12 @@ test_that("can predict on basic vetiver router", {
     preds <- predict(endpoint, mtcars[10:17, 2:3])
     expect_s3_class(preds, "tbl_df")
     expect_equal(nrow(preds), 8)
+    expect_equal(ncol(preds), 1)
+
+    aug <- augment(endpoint, mtcars[10:17, 2:3])
+    expect_s3_class(aug, "tbl_df")
+    expect_equal(nrow(aug), 8)
+    expect_equal(ncol(aug), 3)
 })
 
 test_that("get correct errors", {